3 preset zooms
-Removed continuous zoom updating because it was too taxing on network and system.
-Created 3 preset zoom levels that activate at different camera elevations
-Update buildings to be placed on proper layer

+++ Assets/Scripts/UI/SceneManagerBase.cs
@@ -103,7 +103,7 @@ namespace Assets.Scripts.UI
         public Dictionary<string, SceneDronePort> DronePorts { get; protected set; }
         public Dictionary<string, SceneParkingStructure> ParkingStructures { get; protected set; }
         public Dictionary<string, SceneRestrictionZone> RestrictionZones { get; protected set; }
-        public float CurrentZoom { get; protected set; }
+        public int CurrentZoom { get; protected set; }
 
         #endregion
